[rename] OAuth2 적용을 위한 패키지 분리
